+++ src/main/java/com/abdelhalim/egypt/clinics/api/doctor/repository/DoctorRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DoctorRepository extends JpaRepository<Doctor, Integer>, JpaSpecificationExecutor<Doctor> {
+public interface DoctorRepository extends JpaRepository<Doctor, Long>, JpaSpecificationExecutor<Doctor> {
 
     @Modifying
     @Query("update Doctor d set d.name = ?1 where d.id = ?2")
