+++ src/main/java/com/abdelhalim/egypt/clinics/api/speciality/repository/SpecialtyRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpecialtyRepository extends JpaRepository<Specialty, Integer>, JpaSpecificationExecutor<Specialty> {
+public interface SpecialtyRepository extends JpaRepository<Specialty, Long>, JpaSpecificationExecutor<Specialty> {
 
 
     @Modifying
